refactor: 에디터와 툴바 컴포넌트 분리
