+++ src/app/components/editor/DefaultEditor.module.scss
@@ -1,0 +1,14 @@
+/* Basic editor styles */
+.tiptap {
+  :first-child {
+    margin-top: 0;
+  }
+
+  blockquote {
+    border-left: 3px solid lightgray;
+    margin: 1rem 0;
+    padding-left: 1rem;
+  }
+
+  // TODO heading 등 정의 필요
+}
