+++ src/app/components/editor/DefaultEditor.tsx
@@ -1,7 +1,5 @@
 'use client'
 
-import React, { useState } from 'react'
-
 import Blockquote from '@tiptap/extension-blockquote'
 import Document from '@tiptap/extension-document'
 import Heading from '@tiptap/extension-heading'
@@ -11,7 +9,9 @@ import { BubbleMenu, EditorContent, useEditor } from '@tiptap/react'
 
 import Indent from '@extensions/Indent'
 
-import styles from './styles.module.scss'
+import Toolbar from './Toolbar'
+
+import styles from './DefaultEditor.module.scss'
 
 /**
  * TODO
@@ -28,8 +28,6 @@ import styles from './styles.module.scss'
  */
 
 export default function DefaultEditor() {
-  const [isOpen, setIsOpen] = useState(false)
-  const [isIndentOption, setIsIndentOption] = useState(false)
   const editor = useEditor({
     extensions: [
       Document,
@@ -48,87 +46,14 @@ export default function DefaultEditor() {
     `,
   })
 
-  const getTextFormatOption = () => {
-    if (editor?.isActive('blockquote')) {
-      return '인용'
-    } else if (editor?.isActive('heading')) {
-      return '제목'
-    } else {
-      return '본문'
-    }
-  }
-
-  const toggleText = () => {
-    editor?.commands.unsetBlockquote()
-    editor?.commands.setParagraph()
-  }
-
-  const toggleBlockquote = () => {
-    editor?.chain().focus().toggleBlockquote().run()
-  }
-
-  const toggleHeading = () => {
-    editor?.commands.unsetBlockquote()
-    editor?.chain().focus().toggleHeading({ level: 1 }).run()
-  }
-
   if (!editor) {
     return null
   }
 
   return (
     <>
-      <BubbleMenu
-        editor={editor}
-        tippyOptions={{
-          duration: 100,
-          onHidden: () => {
-            setIsIndentOption(false)
-            setIsOpen(false)
-          },
-        }}
-      >
-        <div className={styles['bubble-menu']}>
-          {/* 텍스트 형식 툴바 */}
-          <button onClick={() => setIsOpen(true)}>{getTextFormatOption()}</button>
-
-          {isOpen && (
-            <div className={styles['dropdown-menu']}>
-              <button
-                onClick={toggleText}
-                className={
-                  !editor.isActive('blockquote') && !editor.isActive('heading')
-                    ? `${styles['is-active']}`
-                    : ''
-                }
-              >
-                본문
-              </button>
-              <button
-                onClick={toggleHeading}
-                className={editor.isActive('heading') ? `${styles['is-active']}` : ''}
-              >
-                제목
-              </button>
-              <button
-                onClick={toggleBlockquote}
-                className={editor.isActive('blockquote') ? `${styles['is-active']}` : ''}
-              >
-                인용
-              </button>
-            </div>
-          )}
-
-          {/* 정렬 툴바 */}
-          <button onClick={() => setIsIndentOption(true)}>정렬</button>
-
-          {isIndentOption && (
-            <div className={styles['dropdown-menu']}>
-              <button onClick={() => editor.commands.indent()}>+ 들여쓰기</button>
-              <button onClick={() => editor.commands.outdent()}>- 내어쓰기</button>
-            </div>
-          )}
-        </div>
+      <BubbleMenu editor={editor} tippyOptions={{ duration: 100 }}>
+        <Toolbar editor={editor} />
       </BubbleMenu>
       <EditorContent editor={editor} className={styles.tiptap} />
     </>
