+++ src/app/components/editor/Toolbar.module.scss
@@ -1,19 +1,5 @@
-/* Basic editor styles */
-.tiptap {
-  :first-child {
-    margin-top: 0;
-  }
-
-  blockquote {
-    border-left: 3px solid lightgray;
-    margin: 1rem 0;
-    padding-left: 1rem;
-  }
-}
-
 /* Bubble menu */
 .bubble-menu {
-  position: relative;
   height: 40px;
 
   display: inline-flex;
